requirements.txt
appdirs==1.4.4
asgiref==3.8.1
tomicwrites==1.4.0
Django==5.1.4
djangorestframework==3.15.2
google-cloud==0.34.0
gunicorn==23.0.0
jsonpointer==2.1
protobuf==3.20.3
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
setuptools==69.5.1
sqlparse==0.5.3
wheel==0.43.0
google-cloud-pubsub
flask
